LOADUI-873 Made some improvements to TableLog Test

+++ loadui-fx-integration-tests/src/main/java/com/eviware/loadui/test/ui/fx/FxIntegrationBase.java
@@ -84,9 +84,9 @@ public class FxIntegrationBase
 			this.knob = knob;
 		}
 
-		public KnobHandle to( int value )
+		public KnobHandle to( long value )
 		{
-			controller.doubleClick( knob ).type( Integer.toString( value ) ).type( KeyCode.ENTER );
+			controller.doubleClick( knob ).type( Long.toString( value ) ).type( KeyCode.ENTER );
 			return this;
 		}
